rename and reorg section stuff
