+++ examples/debug_perf.rs
@@ -3,7 +3,7 @@
 use color_eyre::{
     eyre::Report,
     eyre::{eyre, WrapErr},
-    Help,
+    Section,
 };
 use tracing::instrument;
