+++ src/lib.rs
@@ -370,8 +370,9 @@ pub use eyre::Report;
 pub use eyre::Result;
 use once_cell::sync::OnceCell;
 use section::help::HelpInfo;
-pub use section::{Section, SectionExt};
-use std::fmt::Display;
+#[doc(hidden)]
+pub use section::Section as Help;
+pub use section::{IndentedSection, Section, SectionExt};
 #[cfg(feature = "capture-spantrace")]
 use tracing_error::SpanTrace;
 #[doc(hidden)]
@@ -385,6 +386,8 @@ mod writers;
 
 /// A custom handler type for [`eyre::Report`] which provides colorful error
 /// reports and [`tracing-error`] support.
+///
+/// # Details
 ///
 /// This type is not intended to be used directly, prefer using it via the
 /// [`color_eyre::Report`] and [`color_eyre::Result`] type aliases.
@@ -403,206 +406,14 @@ pub struct Handler {
 
 static CONFIG: OnceCell<config::PanicHook> = OnceCell::new();
 
-/// A helper trait for attaching informational sections to error reports to be
-/// displayed after the chain of errors
-///
-/// `color_eyre` provides two types of help text that can be attached to error reports: custom
-/// sections and pre-configured sections. Custom sections are added via the `section` and
-/// `with_section` methods, and give maximum control over formatting. For more details check out
-/// the docs for [`Section`].
-///
-/// The pre-configured sections are provided via `suggestion`, `warning`, and `note`. These
-/// sections are displayed after all other sections with no extra newlines between subsequent Help
-/// sections. They consist only of a header portion and are prepended with a colored string
-/// indicating the kind of section, e.g. `Note: This might have failed due to ..."
-///
-/// [`Section`]: struct.Section.html
-pub trait Help<T>: private::Sealed {
-    /// Add a section to an error report, to be displayed after the chain of errors.
-    ///
-    /// Sections are displayed in the order they are added to the error report. They are displayed
-    /// immediately after the `Error:` section and before the `SpanTrace` and `Backtrace` sections.
-    /// They consist of a header and an optional body. The body of the section is indented by
-    /// default.
-    ///
-    /// # Examples
-    ///
-    /// ```rust,should_panic
-    /// use color_eyre::{eyre::eyre, eyre::Report, Help};
-    ///
-    /// Err(eyre!("command failed"))
-    ///     .section("Please report bugs to https://real.url/bugs")?;
-    /// # Ok::<_, Report>(())
-    /// ```
-    fn section<D>(self, section: D) -> eyre::Result<T>
-    where
-        D: Display + Send + Sync + 'static;
-
-    /// Add a Section to an error report, to be displayed after the chain of errors. The closure to
-    /// create the Section is lazily evaluated only in the case of an error.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use color_eyre::{eyre::eyre, eyre::Report, Help, SectionExt};
-    ///
-    /// let output = std::process::Command::new("ls")
-    ///     .output()?;
-    ///
-    /// let output = if !output.status.success() {
-    ///     let stderr = String::from_utf8_lossy(&output.stderr);
-    ///     Err(eyre!("cmd exited with non-zero status code"))
-    ///         .with_section(move || stderr.trim().to_string().header("Stderr:"))?
-    /// } else {
-    ///     String::from_utf8_lossy(&output.stdout)
-    /// };
-    ///
-    /// println!("{}", output);
-    /// # Ok::<_, Report>(())
-    /// ```
-    fn with_section<D, F>(self, section: F) -> eyre::Result<T>
-    where
-        D: Display + Send + Sync + 'static,
-        F: FnOnce() -> D;
-
-    /// Add an error section to an error report, to be displayed after the primary error message
-    /// section.
-    ///
-    /// # Examples
-    ///
-    /// ```rust,should_panic
-    /// use color_eyre::{eyre::eyre, eyre::Report, Help};
-    /// use thiserror::Error;
-    ///
-    /// #[derive(Debug, Error)]
-    /// #[error("{0}")]
-    /// struct StrError(&'static str);
-    ///
-    /// Err(eyre!("command failed"))
-    ///     .error(StrError("got one error"))
-    ///     .error(StrError("got a second error"))?;
-    /// # Ok::<_, Report>(())
-    /// ```
-    fn error<E>(self, error: E) -> eyre::Result<T>
-    where
-        E: std::error::Error + Send + Sync + 'static;
-
-    /// Add an error section to an error report, to be displayed after the primary error message
-    /// section. The closure to create the Section is lazily evaluated only in the case of an error.
-    ///
-    /// # Examples
-    ///
-    /// ```rust,should_panic
-    /// use color_eyre::{eyre::eyre, eyre::Report, Help};
-    /// use thiserror::Error;
-    ///
-    /// #[derive(Debug, Error)]
-    /// #[error("{0}")]
-    /// struct StringError(String);
-    ///
-    /// Err(eyre!("command failed"))
-    ///     .with_error(|| StringError("got one error".into()))
-    ///     .with_error(|| StringError("got a second error".into()))?;
-    /// # Ok::<_, Report>(())
-    /// ```
-    fn with_error<E, F>(self, error: F) -> eyre::Result<T>
-    where
-        F: FnOnce() -> E,
-        E: std::error::Error + Send + Sync + 'static;
-
-    /// Add a Note to an error report, to be displayed after the chain of errors.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # use std::{error::Error, fmt::{self, Display}};
-    /// # use color_eyre::eyre::Result;
-    /// # #[derive(Debug)]
-    /// # struct FakeErr;
-    /// # impl Display for FakeErr {
-    /// #     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-    /// #         write!(f, "FakeErr")
-    /// #     }
-    /// # }
-    /// # impl std::error::Error for FakeErr {}
-    /// # fn main() -> Result<()> {
-    /// # fn fallible_fn() -> Result<(), FakeErr> {
-    /// #       Ok(())
-    /// # }
-    /// use color_eyre::Help as _;
-    ///
-    /// fallible_fn().note("This might have failed due to ...")?;
-    /// # Ok(())
-    /// # }
-    /// ```
-    fn note<D>(self, note: D) -> eyre::Result<T>
-    where
-        D: Display + Send + Sync + 'static;
-
-    /// Add a Note to an error report, to be displayed after the chain of errors. The closure to
-    /// create the Note is lazily evaluated only in the case of an error.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # use std::{error::Error, fmt::{self, Display}};
-    /// # use color_eyre::eyre::Result;
-    /// # #[derive(Debug)]
-    /// # struct FakeErr;
-    /// # impl Display for FakeErr {
-    /// #     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-    /// #         write!(f, "FakeErr")
-    /// #     }
-    /// # }
-    /// # impl std::error::Error for FakeErr {}
-    /// # fn main() -> Result<()> {
-    /// # fn fallible_fn() -> Result<(), FakeErr> {
-    /// #       Ok(())
-    /// # }
-    /// use color_eyre::Help as _;
-    ///
-    /// fallible_fn().with_note(|| {
-    ///         format!("This might have failed due to ... It has failed {} times", 100)
-    ///     })?;
-    /// # Ok(())
-    /// # }
-    /// ```
-    fn with_note<D, F>(self, f: F) -> eyre::Result<T>
-    where
-        D: Display + Send + Sync + 'static,
-        F: FnOnce() -> D;
-
-    /// Add a Warning to an error report, to be displayed after the chain of errors.
-    fn warning<D>(self, warning: D) -> eyre::Result<T>
-    where
-        D: Display + Send + Sync + 'static;
-
-    /// Add a Warning to an error report, to be displayed after the chain of errors. The closure to
-    /// create the Warning is lazily evaluated only in the case of an error.
-    fn with_warning<D, F>(self, f: F) -> eyre::Result<T>
-    where
-        D: Display + Send + Sync + 'static,
-        F: FnOnce() -> D;
-
-    /// Add a Suggestion to an error report, to be displayed after the chain of errors.
-    fn suggestion<D>(self, suggestion: D) -> eyre::Result<T>
-    where
-        D: Display + Send + Sync + 'static;
-
-    /// Add a Suggestion to an error report, to be displayed after the chain of errors. The closure
-    /// to create the Suggestion is lazily evaluated only in the case of an error.
-    fn with_suggestion<D, F>(self, f: F) -> eyre::Result<T>
-    where
-        D: Display + Send + Sync + 'static,
-        F: FnOnce() -> D;
-}
-
 // TODO: remove when / if ansi_term merges these changes upstream
 trait ColorExt {
     fn make_intense(self) -> Self;
 }
 
 /// Install the default panic and error report hooks
+///
+/// # Details
 ///
 /// This function must be called to enable the customization of `eyre::Report`
 /// provided by `color-eyre`. This function should be called early, ideally
@@ -613,7 +424,7 @@ trait ColorExt {
 /// function _must_ be called before any `eyre::Report`s are constructed to
 /// prevent the default handler from being installed.
 ///
-/// Examples
+/// # Examples
 ///
 /// ```rust
 /// use color_eyre::eyre::Result;
